refactor: split test_token_summary_appender.py along concern boundaries
- Port 4 orphaned tests to infra split files (1 format_table → filters,
  3 efficiency table → core)
- Move _make_run_skill_event_with_order_id to shared helpers module,
  delete duplicate definitions from both infra files
- Remove 2 hook-existence tests from core.py (authoritative location is
  now the hooks stub)
- Reduce hooks file to 2-test registration stub (script existence +
  source quality)
- Update hooks/CLAUDE.md and infra/CLAUDE.md to reflect new scope
